Proposed updates for the 'x86' branch.
