+++ INSTALL
@@ -398,7 +398,9 @@ MFEM_USE_SIDRE = YES/NO
 
 MFEM_USE_SIMD = YES/NO 
    Enables the high performance templated classes to use specific intrinsics 
-   instead of the AutoSIMD (config/simd/auto.hpp) classe.
+   instead of the AutoSIMD (config/simd/auto.hpp) class. This option should be
+   combined with suitable compiler options, such as -march=native, to enable
+   optimal vectorization.
 
 MFEM_USE_CONDUIT = YES/NO
    Enables support for converting MFEM Mesh and Grid Function objects to and
