+++ config/simd/m256.hpp
@@ -9,15 +9,24 @@
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
 
-#ifndef MFEM_TEMPLATE_CONFIG_SIMD_M256
-#define MFEM_TEMPLATE_CONFIG_SIMD_M256
+#ifndef MFEM_SIMD_M256_HPP
+#define MFEM_SIMD_M256_HPP
+
+#ifdef __AVX__
 
 #include "../tconfig.hpp"
+#include <x86intrin.h>
+
+namespace mfem
+{
+
+template <typename, int, int> struct AutoSIMD;
 
 template <> struct AutoSIMD<double,4,4>
 {
+   typedef double scalar_type;
    static constexpr int size = 4;
-   static constexpr int align_size = 32;
+   static constexpr int align_size = 4;
 
    union
    {
@@ -243,4 +252,8 @@ AutoSIMD<double,4,4> operator/(const double &e,
    return r;
 }
 
-#endif // MFEM_TEMPLATE_CONFIG_SIMD_M256
+} // namespace mfem
+
+#endif // __AVX__
+
+#endif // MFEM_SIMD_M256_HPP
