+++ config/simd/m512.hpp
@@ -9,15 +9,24 @@
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
 
-#ifndef MFEM_TEMPLATE_CONFIG_SIMD_M512
-#define MFEM_TEMPLATE_CONFIG_SIMD_M512
+#ifndef MFEM_SIMD_M512_HPP
+#define MFEM_SIMD_M512_HPP
+
+#ifdef __AVX512F__
 
 #include "../tconfig.hpp"
+#include <x86intrin.h>
+
+namespace mfem
+{
+
+template <typename, int, int> struct AutoSIMD;
 
 template <> struct AutoSIMD<double,8,8>
 {
+   typedef double scalar_type;
    static constexpr int size = 8;
-   static constexpr int align_size = 64;
+   static constexpr int align_size = 8;
 
    union
    {
@@ -231,4 +240,8 @@ AutoSIMD<double,8,8> operator/(const double &e,
    return r;
 }
 
-#endif // MFEM_TEMPLATE_CONFIG_SIMD_M512
+} // namespace mfem
+
+#endif // __AVX512F__
+
+#endif // MFEM_SIMD_M512_HPP
