+++ config/simd/vsx128.hpp
@@ -9,15 +9,24 @@
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
 
-#ifndef MFEM_TEMPLATE_CONFIG_SIMD_VSX128
-#define MFEM_TEMPLATE_CONFIG_SIMD_VSX128
+#ifndef MFEM_SIMD_VSX128_HPP
+#define MFEM_SIMD_VSX128_HPP
+
+#ifdef __VSX__
 
 #include "../tconfig.hpp"
+#include <altivec.h>
+
+namespace mfem
+{
+
+template <typename,int,int> struct AutoSIMD;
 
 template <> struct AutoSIMD<double,2,2>
 {
+   typedef double scalar_type;
    static constexpr int size = 2;
-   static constexpr int align_size = 16;
+   static constexpr int align_size = 2;
 
    union
    {
@@ -231,4 +240,8 @@ AutoSIMD<double,2,2> operator/(const double &e,
    return r;
 }
 
-#endif // MFEM_TEMPLATE_CONFIG_SIMD_VSX128
+} // namespace mfem
+
+#endif // __VSX__
+
+#endif // MFEM_SIMD_VSX128_HPP
